Move things to their right place
