+++ src/utils/parsePropValue.ts
@@ -1,0 +1,5 @@
+import { stringOrNumber } from '../typings/general';
+
+export const parsePropValue = (value: stringOrNumber) => {
+  return typeof value === 'string' ? value : `${value}rem`;
+};
